@@ -13,31 +13,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/common": "^14.2.10",
+    "@angular/core": "^14.2.10",
+    "@angular/forms": "^14.2.10",
+    "@angular/platform-browser": "^14.2.10",
+    "@angular/platform-browser-dynamic": "^14.2.10",
+    "@angular/router": "^14.2.10",
     "@ionic-native/core": "^5.36.0",
     "@ionic-native/splash-screen": "^5.36.0",
     "@ionic-native/status-bar": "^5.36.0",
-    "@ionic/angular": "^5.9.1",
+    "@ionic/angular": "^6.3.6",
     "@ionic/storage-angular": "^3.0.6",
     "rxjs": "~6.6.7",
     "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler": "~13.3.11",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.9",
+    "@angular/cli": "^14.2.9",
+    "@angular/compiler": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.10",
+    "@angular/language-service": "^14.2.10",
     "@ionic/angular-toolkit": "^5.0.3",
-    "@types/node": "^16.11.9",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^16.11.9",
     "codelyzer": "^6.0.2",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
